logging: add trace statements when changes are ignored
