+++ internal/terraform/node_resource_abstract_instance.go
@@ -1252,7 +1252,7 @@ func (n *NodeAbstractResource) processIgnoreChanges(prior, config cty.Value, sch
 		return config, nil
 	}
 
-	ignoreChanges := traversalsToPaths(n.Config.Managed.IgnoreChanges)
+	ignoreChanges, keys := traversalsToPaths(n.Config.Managed.IgnoreChanges)
 	ignoreAll := n.Config.Managed.IgnoreAllChanges
 
 	if len(ignoreChanges) == 0 && !ignoreAll {
@@ -1260,6 +1260,8 @@ func (n *NodeAbstractResource) processIgnoreChanges(prior, config cty.Value, sch
 	}
 
 	if ignoreAll {
+		log.Printf("[TRACE] processIgnoreChanges: Ignoring all changes for %s", n.Addr)
+
 		// Legacy providers need up to clean up their invalid plans and ensure
 		// no changes are passed though, but that also means making an invalid
 		// config with computed values. In that case we just don't supply a
@@ -1282,6 +1284,7 @@ func (n *NodeAbstractResource) processIgnoreChanges(prior, config cty.Value, sch
 
 		return ret, nil
 	}
+	log.Printf("[TRACE] processIgnoreChanges: Ignoring changes for %s at [%s]", n.Addr, strings.Join(keys, ", "))
 
 	if prior.IsNull() || config.IsNull() {
 		// Ignore changes doesn't apply when we're creating for the first time.
@@ -1296,36 +1299,49 @@ func (n *NodeAbstractResource) processIgnoreChanges(prior, config cty.Value, sch
 
 // Convert the hcl.Traversal values we get form the configuration to the
 // cty.Path values we need to operate on the cty.Values
-func traversalsToPaths(traversals []hcl.Traversal) []cty.Path {
+func traversalsToPaths(traversals []hcl.Traversal) ([]cty.Path, []string) {
 	paths := make([]cty.Path, len(traversals))
+	keys := make([]string, len(traversals))
 	for i, traversal := range traversals {
-		path := traversalToPath(traversal)
+		path, key := traversalToPath(traversal)
 		paths[i] = path
+		keys[i] = key
 	}
-	return paths
+	return paths, keys
 }
 
-func traversalToPath(traversal hcl.Traversal) cty.Path {
+func traversalToPath(traversal hcl.Traversal) (cty.Path, string) {
 	path := make(cty.Path, len(traversal))
+	var key strings.Builder
 	for si, step := range traversal {
 		switch ts := step.(type) {
 		case hcl.TraverseRoot:
 			path[si] = cty.GetAttrStep{
 				Name: ts.Name,
 			}
+			key.WriteString(ts.Name)
 		case hcl.TraverseAttr:
 			path[si] = cty.GetAttrStep{
 				Name: ts.Name,
 			}
+			key.WriteString(".")
+			key.WriteString(ts.Name)
 		case hcl.TraverseIndex:
 			path[si] = cty.IndexStep{
 				Key: ts.Key,
+			}
+			if ts.Key.Type().IsPrimitiveType() {
+				key.WriteString("[")
+				key.WriteString(tfdiags.CompactValueStr(ts.Key))
+				key.WriteString("]")
+			} else {
+				key.WriteString("[...]")
 			}
 		default:
 			panic(fmt.Sprintf("unsupported traversal step %#v", step))
 		}
 	}
-	return path
+	return path, key.String()
 }
 
 func processIgnoreChangesIndividual(prior, config cty.Value, ignoreChangesPath []cty.Path) (cty.Value, tfdiags.Diagnostics) {
